// BackEnd/src/types/enum.types.ts
export enum UserRole {
    ADMIN = "admin",
    USER = "user",
    MODERATOR = "moderator",
}
export enum AttachmentType {
    IMAGE = "image",
    VIDEO = "video",
    FILE = "file",
    AUDIO = "audio",
}
export enum DbFriendStatus {
    PENDING = "pending",
    ACCEPTED = "accepted",
}
export enum UserRelationStatus {
    FRIEND = "friend",
    SENT = "sent",
    RECEIVED = "received",
    NOT_FRIEND = "not_friend",
}